+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/internal/ComposeTypeUtils.kt
@@ -3,6 +3,8 @@ package com.mapbox.maps.extension.compose.style.internal
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import com.mapbox.bindgen.Value
+import com.mapbox.maps.TileCacheBudgetInMegabytes
+import com.mapbox.maps.TileCacheBudgetInTiles
 import com.mapbox.maps.extension.style.utils.ColorUtils
 import com.mapbox.maps.extension.style.utils.TypeUtils
 
@@ -10,6 +12,20 @@ internal object ComposeTypeUtils {
   fun wrapToValue(value: Any): Value {
     return when (value) {
       is Color -> Value(ColorUtils.colorToRgbaString(value.toArgb()))
+      is TileCacheBudgetInMegabytes -> Value(
+        hashMapOf(
+          "megabytes" to Value(
+            value.size
+          )
+        )
+      )
+      is TileCacheBudgetInTiles -> Value(
+        hashMapOf(
+          "tiles" to Value(
+            value.size
+          )
+        )
+      )
       else -> TypeUtils.wrapToValue(value)
     }
   }